Move global exposition out from script loading
